rewrite rendering
Environment now draws everything in one draw call. Materials and
their components are attached to buffers as part of a single shader.
Degenerate triangles are inserted as necessary for rendering correctly.

This also required a rewrite of shadows. The new implementation is
simpler and more friendly to ellipses but the behavior is different
than before. This has also allowed animating material changing shape
by adjusting the roundness of a material where zero is rectangular,
width/2 is an ellipse, and everything in between is rounded rectangle.
Additional work on shadows will be required to improve aesthetics.

Overall, a sample app that previously ran at 20 FPS is now running
at 60 FPS.

Fixes #4