text: support variable width, enumerating glyphs, tight packing
Overhaul of the text package pushing positioning into material
to achieve tighter packing in the texture. Supports variable
width fonts and a simpler way to generate textures that
enumerate all possible glyphs in a font file.

Positioning is done based on normalized unit values and isn't
expected to be exact, only good enough for real-time rendering
on low-end mobile hardware.

The Droid Sans sample for example/text has all available glyphs
with room to spare for potentially packing smaller versions of
the glyphs in the future.

There is a reduction in quality of lower point sizes compared
to previously and in all sizes on mobile, but this should be
disregarded as those rendering values were tuned for a specific
instance and should go away in the future.

Fixes #5